Unify Entry + Request in pkg/beacon/relay/entry
We had a pkg/beacon/entry package with a bunch of extraneous code and a
Request object, that in turn referenced pkg/beacon/relay.Entry, which
was being referenced everywhere else.

To avoid import cycles and simplify the package structure a bit, we
create a new pkg/beacon/relay/entry package that contains both Entry and
Request, and use that throughout the codebase.