Fix gitignore for pkg/chain/gen
There was a subdirectory gitignore that was ignoring all .go files in
pkg/chain, but this included the generator code itself. All generated
code goes into subdirectories, and there is a rule that handles this
kind of ignoring for pkg/gen from when the async generator was added.
That rule is now generalized to all gen/ subpackages, and the
subdirectory gitignore is gone.