ErrorResolver now carries From information
When resolving errors, ErrorResolver has to reissue a contract call to
retrieve the return data and extract the error from it. Without the
correct From, msg.sender is not set correctly for the contract methods,
so the returned error is returned under different circumstances than
when the original transaction failed. Whoops!