Add global transaction submission mutex
When transactions are submitted, they require a valid nonce. The nonce
is equal to the count of transactions the account has submitted so far,
and for a transaction to be accepted it should be monotonically greater
than any previous submitted transaction. To do this, transaction
submission asks the Ethereum client it is connected to for the next
pending nonce, and uses that value for the transaction. Unfortunately,
if multiple transactions are submitted in short order, they may all get
the same nonce.

Currently, transaction submission does not require particularly high
throughput. As such, a new global transaction mutex acts to serialize
all transaction submissions, so that each transaction gets the
next valid nonce (which could be the same as a previous one if a
previous one fails gas estimation, for example).

This may require adjusting if higher throughputs end up being necessary.