Add Call<method> for non-const contract methods
Call<method> generates a non-mutable call to the Ethereum node that
essentially acts as a dry-run of the mutating transaction, without
charging any gas or subtracting any sent value from the sender's account
balance, and without signing or submitting the transaction to the
transaction pool.

It can also be evaluated at a particular block.