Add CLI subcommands for non-const contract methods
These take the --submit parameter to allow for a proper transaction
submission to occur, and default to doing a non-mutating dry-run call of
the non-const method and returning that result. They also accept --value
to allow sending ETH alongside the transaction, and, like const methods,
--block to do dry-run calls against a particular block.

When submitting the transaction, the output is the transaction hash.
When calling it, the output is the return value, or nothing if the
non-const method has no return value.