Add transaction logging to generated contracts
Currently, only transactions are logged; that is, simple calls are not
logged. Transactions log the method that is being invoked, the values
being passed to it, any ETH value that is included, and, if the
transaction submits successfully, the hash of the resulting transaction.

Notably, each contract gets its own logger. This logger starts with
keep-contract-, followed by the name of the contract class. This allows
for setting the log level on all contracts with `keep-contract-*`, as
well as setting a per-contract log level with e.g.
`keep-contract-KeepRandomBeaconService`.