Add README for keep-common repo
The repo currently holds all prior history for the code that is known to
merge into common code. Future code that is split into this repository
may or may not be pulled in with history.

The README includes the goal directory structure, though the current
structure is a direct lift from the previous keep-core repo structure.