Add promise template generation and generated file
Exceptionally, for generated template variables, the generated file is
committed to git alongside the directive to generate it. The generation
also occurs outside of a `gen` package. This allows for `go install` to
make the full executable with template available to any package that is
trying to use the code generator.