Rework promise generator to be configurable
Until now, the promises to generate have been hardcoded in the generator
executable, as has the output directory. For usage by external packages,
that needs to be configurable.

The flag package is now used to allow specifying an output directory,
which still defaults to `./async`. The command line additionally takes a
set of Go type names, which can include package information. When
package information is included, it is placed in the template's import
block; when it isn't, it is inferred during import organization.

The code for parsing a Go type given to the executable into a promise
configuration (which includes the actual type the promise references,
the package the type will be in, the promise struct's prefix, and the
filename for the promise) lives in its own file for easier analysis.

As one example,
*github.com/keep-network/keep-core/pkg/relay/event.GroupRegistration
will use *event.GroupRegistration as the promise type,
EventGroupRegistrationPromise as the promise struct name, and will
explicitly import github.com/keep-network/keep-core/pkg/relay/event in
the generated Go file. That file will be named
event_group_registration_promise.go

Conversely, *big.Int will use *big.Int as the promise type,
BigIntPromise as the promise struct name, and will let `big` be resolved
to an import path by the Go compiler during import organization (which
happens as a part of code generation). The file will be named
big_int_promise.go.

In the process, the promise generator has been adapted to use the
errorAndExit pattern used in the template generator.

Finally, the BigInt test removed in previous commits is now back, but
used exclusively for testing the generation process. The test verifies
that a promise is properly generated and works correctly.