Add Ethereum config struct to pkg/chain/ethereum
This was being tracked on a per-application basis, but as generators are
added to this repo that will interact with Ethereum, they need a way to
express their expected way of consuming Ethereum configuration
information. As such, the config structure for Ethereum configuration
specifically is moved to pkg/chain/ethereum in this common repository.