Add ConfigFunc contract parameter and related code
The ConfigFunc parameter specifies a function that, given a
configuration file name, returns a pkg/chain/ethereum.Config struct
corresponding to that file, or an error if an Ethereum config couldn't
be parsed. By default, it's set to config.ReadEthereumConfig, though
resolution of the `config` package is left to Go's import resolution
during code generation.

This decouples the generated contract code from a particular containing
configuration structure or parsing function, focusing on the Ethereum
config structure specifically.