Compile Ethereum binding templates into generator
The Ethereum contract binding generator now uses the template generator
helper to generate .go files from .go.tmpl files, allowing the templates
to be compiled directly into the binding generator. This means the
template generator executable can be invoked directly without having to
somehow provide the templates out of band.

Note also that the generated files are checked in, unlike in many
code-generation cases, so that a dep or go.mod dependency on this
generator need not go through a build process to have a working
executable. Instead, a simple `go run` or `go build` will be sufficient.