Output function name with the first letter lowercase from camelCase
We introduced `toCamelCase` to use the same function name normalization
strategy as go-ethereum and support functions like `__beaconCallback`.

This change broke `ethereum` command. The subcommands for ethereum methods
were being prefixed by `-`. urfave/cli, the library we use for CLI
definition and parsing, does not fail if subcommands prefixed by - are used.
However, it does fail to recognize them; instead, it tries to recognize them
as flags, and flags with that name have not been defined.

`toCamelCase` used to uppercase the first letter. Later, in
`buildMethodInfo` where we create `dashedName` we were prefixing the
function name with `-` with uppercase-first letter.

The fix is easy, we need to ensure the first letter stays lowercase during
the normalization. Later, in `buildMethodInfo`, we specify explicitly where
do we want lowercase and uppercase with `uppercaseFirst` and `lowercaseFirst`.