Improved error handling in moveAll function
We are now explicit that we os.Rename only if the target directory does
not exist (os.IsNotExist(err)) and we don't do that if it's any other
type of an error. In such case, we return from the function doing
nothing.

We also no longer ignore errors from ioutil.ReadDir. If the error
happens, we return the function doing nothing.