Local nonce manager
Nonce manager is used by the generated contracts to track the current
nonce and update it after each successfully submitted transaction.
Tracking the nonce locally is required when transactions are submitted
from multiple goroutines for when Ethereum clients are deployed behind a
load balancer and their mempools are not always in sync.