Parse ether value from a text
In a configuration file we are providing value for MaxGasPrice
and BalanceAlertThreshold. If we provide it in wei we are limited to
around 9.22 ether. To be more flexible and allow providing greater value
especially for balance monitoring we added a possibility to provide the
value as string with three types of units: wei, Gwei and ether.
We are also supporting decimals.

Few exaples of how the value can be provided:
- `BalanceAlertThreshold = 100000000000`
- `BalanceAlertThreshold = "100 Gwei"`
- `BalanceAlertThreshold = "2 ether"`
- `BalanceAlertThreshold = "3.4 ether"`