Accept subscriptionFailed and thresholdViolated in WithResubscription
The mechanism monitors the time elapsed between resubscription attempts and
if it is shorter than the specificed alertThreshold, it writes the time
elapsed between resubscription attempts to the thresholdViolated channel to
alarm about potential problems with the stability of the subscription.

Every error returned by the wrapped subscription function, is written to
subscriptionFailed channel.

If the calling code is interested in reading from thresholdViolated and/or
subscriptionFailed channel, appropriate readers need to be set up *before*
WithResubscription is called.

Writes to thresholdViolated and subscriptionFailed channels are non-blocking
and are not stopping resubscription attempts.

Those two channels will let the client code log warning/errors or even
kill the subscription in case alrt threshold is violated too many times.