OnEvent and PipeTo event subscription
The new event subscription mechanism allows to install event callback
handler function using OnEvent function as well as pass a channel to
which events should be piped.

So far all our event subscriptions use function handlers. While it is
convenient in some cases, in some other cases it is the opposite. For
example, OnBondedECDSAKeepCreated handler in ECDSA client works
perfectly fine as a function. It triggers the protocol and does not have
to constantly monior the state of the chain. On the other hand,
OnDKGResultSubmitted from beacon client needs to monitor the chain and
exit the process of event publication in case another node has already
done it. In this case, the code could be better structured with
channel-based subscription.

Another important change introduced in this commit is a background event
check loop that is calling GetPastEvents every 15 minutes (by default).
This logic was added to make sure no events are lost from the client's
perspective. We observed on mainnet that some third party Ethereum
providers are often not delivering events. It may also happen that the
event was emitted when client was attempting to reconnect to Ethereum.
Both those problems should be fixed by the background loop calling
getPastEvents next to the regular event subscription.

Important: with this solution, client will receive even more event
duplicates than before. It is important to implement an appropriate
duplicate filtering logic in the handler.