resubscribeFn renamed to subscribeFn
This function was called resubscribeFn just because of
event.ResubscribeFn type. However, this function is called for the
first-ever subscription as well and we do have subscriptionFailedFn as
another parameter, named this way and not resubscriptionFailedFn for
exactly the same reason - it can be called for the first-ever
subscription. Moreover, each resubscription is in fact another
subscription attempt.