Create `ethlike` package
Here we create the `ethlike` package which will
eventually contain all types common for ETH-like
chains. The intention behind that change is to
decouple the auxiliary tools from the Ethereum
implementation and make them ready to work with
all ETH-like chains, for example, Celo.

The first application of the `ethlike` package
is the refactoring of the `MiningWaiter` tool also
included in this commit. Apart from that,
the `ethliketest` package has been created. It
contains mock implementations of all `ethlike`
types and it's intended to make testing easier.