Revert changes made around client wrappers
This change takes a step back regarding client
wrappers (logging wrapper and rate limiter). The
cause of that decision is the fact that a wrapped
`ethlike.Client` can't be directly used by
generated contract bindings. Those bindings
are inseparably bound with the
`go-ethereum` package so a forcible introduction
of `ethlike` types will have a unacceptable
negative impact on code quality.

However, this change leaves the core rate limiter
logic in the `ethlikeutil` package in order to
make it reusable by other chains implementations.