Final cleanup in the `ethlike` package
Revisited the `ethlike` package content and got
rid of the general `ethlike/interface.go` file.
Instead several small interfaces have been added
in places of their actual usage. Such a design
is cleaner and closer to the idiomatic way of
using interfaces in Go.

Also, removed the `ethliketest` package in favor
of local mock implementations which are placed
next to the tests which leverages them.