Improve block handling in the block source adapter
Stopping intermediary goroutine once the context
is done is not a proper behavior. This is because
the context passed in function params is intended
to control the creation of the subscription, not
their lifetime. This can cause a premature
stop of the goroutine as the context is
often set to a short deadline. Because of that, a
separate stop channel should be used instead to
stop the intermediary goroutine once the
unsubscribe function is called.