Error resolver for Celo
Created an error resolver for Celo. For now it is
a duplicate of the Ethereum error resolver.
There is no simple way to make it common
and move it to the `ethlike` package. There are
several specific `go-ethereum` types and trying
to hide them behind interfaces is hard. The amount
of required work doesn't justify potential
benefits.