Duplicate resubscription mechanism
The `WithResubscription` function from `ethutil`
package should be also duplicated in `celoutil`
package. It strongly depends on `go-ethereum`
module and it's short. This weaken the reasons
about making it an `ethlike` component. Making
it common seems to be an overkill as it would also
require dedicated chain adapters which will make
this simple code cumbersome.

Regarding the auxiliary `SubscribeOpts` type,
it also makes sense to duplicate them instead
of making it common. Subscription params may
differ between chains as well as their default
values.