Create GH Actions workflow for build&test of Go client
As described in RFC-18, there is a need for a refactorization of Keep
and tBTC release processes in order to reduce human involvment and make
the processes faster and less error prone. A part of this task is
migration from CircleCI jobs to GitHub Actions. This commit creates a
GitHub Action workfow for building and testing of Go client.

At this point, solution utilizes EnricoMi/publish-unit-test-result-action@v1
action for publishing of the test results, but this may be replaced in the
future by use of other action or use of problem matchers.