Retry balance checks on failures
We want to use a retry mechanism for single executions of the balance check.
The retries will be performed during the period of `retryTimeout`,
logging a warning on each error. When the timeout is hit an error
will be logged and retries stopped. A next balance check will be triggered at the tick.