Refactor mining waiter to be ready for EIP-1559
EIP-1559 introduced a new type of transactions
which no longer use the `gasPrice` field and
rely on `gasFeeCap` and `gasTipCap` parameters.
The mining waiter can't support the dynamic
fee txs out of the box so we need to extend it.
To make the extension possible, we need to
refactor the mining waiter structure.