Handle structured outputs
Solidity method can return multiple outputs. If the outputs are named
they are gathered into a structured output (a struct). If at least one
outputs is unnamed it is considered an unstructured output and code
generated by abigen return multiple outputs that we need to handle and
pack into a struct.