Modify reading ethereum config in the command
We need to intialize a connection to the Ethereum chain as part of the
contract handle initialization.

In V1 we were using `config.GetEthereumConfig` function with pasing a
path to the config file.

In V2 we added support of cli flags for configuring properties.

The previous solution had difficulties due to the complexity of
different levels at which the flags were defined for commands. From the
generated code perspective we don't want to go much into the commands
structure and implemantation. We expect the parten command to expose the
config for initialization.