Improve type parsing function discovery
For each of the params we need to find out which function should be used
to parse a string argument to a type expected by the contract's Go
binding.

If any of the method's (function's) input params type is not associated with
any parsing function the function is not callable as a command (see
`default` path in the `switch` statement). This caused a lot of
functions not being exposed in the generated ethereum CLI commands.

In this commit we want to support more types, to have more functions
exposed in the CLI.

We add support of:
- bool
- uint* with bit size greater or equal 64

To be able to use `strconv.ParseUint` we had to introduce the `parsingFn` as a format for `printf` function, where `%s` will be replaced with an argument.