Support tuples (structures) in ethereum commands
Some of the contract's methods accept tuples (structures) as arguments.
We want to support such functions and generate ethereum CLI commands for
them.

When a param is a tuple we expect that it's string value in the CLI
execution be provided as a JSON. In the generated code we will parse the
JSON to a Go type defined for the structure.

We add `_json` suffix to such parameters to be printed in the CLI help
and add some context what is expected to be provided when calling it.