Merge diagnostics and metrics into a single clientinfo registry
Previously, we were exposing diagnostics and metrics on separate ports but the
reality was that we had two HTTP servers exposing exactly the same content: the
diagnostics port could be queried for both /diagnostics and /metrics and the
metrics port could be queried for both /metrics and /diagnostics. This change
simplifies the setup into a single clientinfo package and registry exposing
metrics and diagnostics.