Support bytes20 and bytes32 in Ethereum contract bindings
There are quite a few functions in our Solidity contracts that accept `bytes20`
or `bytes32` as a parameter. One of many examples are
`requestHeartbeat(bytes20 walletPubKeyHash, bytes calldata message)` from the
`WalletCoordinator` contract and `getWallet(bytes32 walletID)` from the
`WalletRegistry` contract. Given there was no parsing function defined in our
Ethereum contract parsing logic, all functions accepting `bytes32` and `bytes20`
were skipped during the code generation:

```
WARNING: Unsupported param type for method getWallet:
  ABI Type: bytes32
  Go Type:  [32]byte
  the method won't be callable with 'ethereum' command
```

This changeset adds two parsing functions: one for `bytes20` and another for
`bytes32` Solidity type. In theory, we could have other variations of `bytesN`
but this is not the case for our codebase, and defining a generic function is
more complicated given we do not deal with slices but fixed-size arrays. Even
though the proposed solution does not work for all `bytesN` types, it is enough
for now.